Fix Bun JSON output startup path (Fixes #2240)
